requirements.txt
langchain
langchain-community
openai
fastapi
uvicorn
python-dotenv
redis
celery
fastapi-mail
sentence-transformers
langchain_openai
weasyprint
beautifulsoup4
rank_bm25
faiss-cpu
pathlib